+++ litebox/src/broker/event_counter.rs
@@ -9,31 +9,15 @@ use litebox_broker_protocol::{
     WaitEventRequest, WaitOutcome,
 };
 
+use super::{BrokerControl, BrokerState};
 use crate::{
     event::{
         Events, IOPollable, observer::Observer, polling::Pollee, polling::TryOpError,
         wait::WaitContext,
     },
     platform::TimeProvider,
-    sync::{RawSyncPrimitivesProvider, RwLock},
+    sync::RawSyncPrimitivesProvider,
 };
-
-/// Error returned by the deployment-provided broker control path.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct BrokerControlError;
-
-/// Local-core access to the negotiated broker control channel.
-///
-/// LiteBox owns broker-backed local objects and constructs broker protocol
-/// requests. Deployment code owns endpoint selection and supplies the connected
-/// transport behind this protocol-level boundary.
-pub trait BrokerControl: Send + Sync {
-    /// Sends one active broker request and returns its response.
-    fn request(
-        &self,
-        request: BrokerRequest,
-    ) -> core::result::Result<BrokerResponse, BrokerControlError>;
-}
 
 /// Errors returned by broker-backed local-core event counters.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -64,21 +48,7 @@ pub struct EventCounter<Platform: RawSyncPrimitivesProvider + TimeProvider> {
     pollee: Pollee<Platform>,
 }
 
-pub(crate) struct BrokerState<Platform: RawSyncPrimitivesProvider> {
-    control: RwLock<Platform, Option<Arc<dyn BrokerControl>>>,
-}
-
 impl<Platform: RawSyncPrimitivesProvider> BrokerState<Platform> {
-    pub(crate) fn new() -> Self {
-        Self {
-            control: RwLock::new(None),
-        }
-    }
-
-    pub(crate) fn set_control(&self, broker_control: Arc<dyn BrokerControl>) {
-        *self.control.write() = Some(broker_control);
-    }
-
     pub(crate) fn create_event_counter(
         &self,
         initial_count: u64,
